requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-js-asset==2.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
isort==5.10.1
Pillow==9.2.0
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.5
sqlparse==0.4.3
termcolor-whl==1.1.2
tzdata==2022.5
yapf==0.32.0
yaspin==2.2.0
